Prettier

+++ src/components/views/Team/TeamProfile.js
@@ -1,30 +1,25 @@
 import { useHistory } from "react-router-dom";
-import {
-  api,
-  handleError
-} from "../../../helpers/api";
+import { api, handleError } from "../../../helpers/api";
 import React, { useEffect, useState } from "react";
 import { Spinner } from "../../ui/Spinner";
 import BaseContainer from "../../ui/BaseContainer";
 import { Button } from "../../ui/Button";
 import avatar from "../../../images/avatar1.png";
 import "styles/views/TeamProfil.scss";
-import {StyledDialog} from "../../ui/StyledDialog";
+import { StyledDialog } from "../../ui/StyledDialog";
 
 //component for a TEAM MEMBER
 export const TeamMember = ({ teamMember, onClick }) => (
   <div className="team member container2">
     <div className="team member icon">
-      <img src={avatar}  alt={" "}/>
+      <img src={avatar} alt={" "} />
     </div>
     <div className="team member username">{teamMember.user.username}</div>
     <div className="team member email">{teamMember.user.email}</div>
     {teamMember.isAdmin ? <div className="team member admin">Admin</div> : ""}
-    {sessionStorage.getItem("isAdmin")==="true" && teamMember.isAdmin === false ? (
-      <div
-        className="team member removebutton "
-        onClick={onClick}
-      >
+    {sessionStorage.getItem("isAdmin") === "true" &&
+    teamMember.isAdmin === false ? (
+      <div className="team member removebutton " onClick={onClick}>
         Remove
       </div>
     ) : (
@@ -68,12 +63,9 @@ export const TeamProfile = () => {
 
   async function removeUser(userId) {
     try {
-      await api.delete(
-        `/teams/${team.id}/users/${userId}`,
-        {
-          headers: { token: sessionStorage.getItem("token") },
-        }
-      );
+      await api.delete(`/teams/${team.id}/users/${userId}`, {
+        headers: { token: sessionStorage.getItem("token") },
+      });
     } catch (error) {
       alert(
         `Something went wrong with deleting the user from the team: \n${handleError(
@@ -81,7 +73,7 @@ export const TeamProfile = () => {
         )}`
       );
     }
-      window.location.reload();
+    window.location.reload();
   }
 
   async function handleDeleteTeam() {
@@ -164,7 +156,7 @@ export const TeamProfile = () => {
           </div>
         ) : null}
       </div>
-      <div/>
+      <div />
       {content}
     </BaseContainer>
   );
